+++ test/integration/repo_install/install_test.rb
@@ -1,0 +1,3 @@
+describe package('nginx') do
+  it { should be_installed }
+end