+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of the nginx cookbook.
 
 ## Unreleased
 
+- Add `repo_train` property to nginx_install to select stable/mainline when installing from the nginx repo.
+- Add `packages_versions` property to nginx_install to specify specific package versions.
+
 ## 11.0.0 - *2020-12-02*
 
 - Refactor cookbook to remove nxensite/nxdissite